[SNAPPYDATA] reverting changes to increase DECIMAL precision to 127
The changes to DECIMAL precision were incomplete and broken in more ways than one.
The other reason being that future DECIMAL optimization for operations in
generated code will depend on value to fit in two longs and there does not seem
to be a practical use-case of having precision >38 (which is not supported
    by most mainstream databases either)

Renamed UnsafeRow.todata to toData for consistency.

Conflicts:
	sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/codegen/UnsafeArrayWriter.java

Conflicts:
	sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/UnsafeRow.java
	sql/catalyst/src/main/scala/org/apache/spark/sql/types/DecimalType.scala