Add message to redirect PRs upstream if possible (#607)
* Add message to redirect PRs upstream if possible

We want to re-direct community dev upstream as much as possible. However, some contributions impact components (e.g. shuffle server) that do not yet exist upstream. To handle this, we decided to add this message and leave it up to developers, but encourage them to submit upstream unless it isn't feasible.

* Add dev mailling list and jira links