[SNAP-2225] Removed OrderlessHashPartitioning. (#95)
Handled join order in optimization phase. Also removed custom changes in HashPartition. We won't store bucket information in HashPartitioning. Instead based on the flag "linkPartitionToBucket" we can determine the number of partitions to be either numBuckets or num cores assigned to the executor.
Reverted changes related to numBuckets in Snappy Spark.
Conflicts:
	sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/physical/partitioning.scala
	sql/core/src/main/scala/org/apache/spark/sql/execution/exchange/EnsureRequirements.scala
	sql/core/src/main/scala/org/apache/spark/sql/execution/exchange/ShuffleExchangeExec.scala