[SDENT-62] Making method synchronized as it is being invoked concurrently by two (#175)
Making the method synchronized as it is being invoked concurrently by two
threads while shutting down the executor which sometimes results in
one of the thread failing with IOException as the file is already
deleted by the other thread. Although this doesn't leave any orphan
files, the exception is logged on the logs causing confusion.