[SNAP-3033] - Fixing a failing test in snappy compatibility suite (#178)
The reason behind the failure was that one of the private fields of the
`StreamingQueryManager` class was being accessed using scala test's
`PrivateMethod`. However, `SnappyStreamingQueryManager` introduced as
part of SNAP-3033 does not contain the same private field hence we are
accessing the same field from the super class using reflection.